ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
groq==0.9.0
firebase-admin==6.4.0
httpx==0.27.0
